requirements.txt
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-reflection==1.80.0
grpcio-status==1.80.0
grpcio-tools==1.80.0
protobuf==6.33.6
psycopg2-binary==2.9.12
setuptools==82.0.1
typing_extensions==4.15.0
